{"schema":"decodex/commit/1","summary":"Move pairwise shift support","authority":"manual"}
